pyproject.toml
[project]
name = "sozo"
version = "0.1.0"
license = "MIT"
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
   "dlib>=19.24.6",
   "fastapi>=0.115.4",
   "imutils>=0.5.4",
   "jinja2>=3.1.4",
   "msgspec>=0.18.6",
   "numpy>=2.1.3",
   "opencv-python>=4.10.0.84",
   "pyzmq>=26.2.0",
   "uvicorn[standard]>=0.32.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["apps", "libs", "scripts", "tools"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
ignore = ["B905", "D417", "E501", "PLR0913"]

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_missing_imports = true
